+++ service/controller/resource/clustersecret/desired.go
@@ -11,7 +11,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha4"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/k8smetadata/pkg/annotation"
@@ -20,9 +19,6 @@ import (
 
 	capi "sigs.k8s.io/cluster-api/api/v1alpha4"
 
-	capvcd "github.com/giantswarm/cluster-apps-operator/v2/api/capvcd/v1beta1"
-
-	capo "github.com/giantswarm/cluster-apps-operator/v2/api/capo/v1alpha4"
 	"github.com/giantswarm/cluster-apps-operator/v2/pkg/project"
 	"github.com/giantswarm/cluster-apps-operator/v2/service/controller/key"
 )
@@ -55,36 +51,6 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) ([]*cor
 	}
 
 	values := map[string]interface{}{}
-
-	{
-		infrastructureRef := cr.Spec.InfrastructureRef
-		if infrastructureRef != nil {
-			switch infrastructureRef.Kind {
-			case "OpenStackCluster":
-				var infraCluster capo.OpenStackCluster
-				err = r.k8sClient.CtrlClient().Get(ctx, client.ObjectKey{Namespace: infrastructureRef.Namespace, Name: infrastructureRef.Name}, &infraCluster)
-				if err != nil {
-					return nil, microerror.Mask(err)
-				}
-
-				values["cloudConfig"], err = r.generateOpenStackCloudConfig(ctx, infraCluster)
-				if err != nil {
-					return nil, microerror.Mask(err)
-				}
-			case "VCDCluster":
-				var infraCluster capvcd.VCDCluster
-				err = r.k8sClient.CtrlClient().Get(ctx, client.ObjectKey{Namespace: infrastructureRef.Namespace, Name: infrastructureRef.Name}, &infraCluster)
-				if err != nil {
-					return nil, microerror.Mask(err)
-				}
-
-				values["global"], err = r.generateCloudDirectorConfig(ctx, infraCluster)
-				if err != nil {
-					return nil, microerror.Mask(err)
-				}
-			}
-		}
-	}
 
 	secretSpecs := []secretSpec{
 		{