+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.3] - 2023-03-09
+
 ### Changed
 
 - Enable private clusters for cloud-director, openstack and vsphere.
@@ -344,7 +346,8 @@ cluster.
 
 - Initial version based on app related logic extracted from cluster-operator.
 
-[Unreleased]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.8.2...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.8.3...HEAD
+[2.8.3]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.8.2...v2.8.3
 [2.8.2]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.8.1...v2.8.2
 [2.8.1]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.8.0...v2.8.1
 [2.8.0]: https://github.com/giantswarm/cluster-apps-operator/compare/v2.7.0...v2.8.0